requirements.txt
diff-match-patch==20121119
Django==1.11.29
django-ckeditor==5.4.0
django-disqus==0.5
django-el-pagination==3.1.0
django-googlesearch==0.2
django-import-export==0.4.5
django-jet==1.0.1
django-js-asset==1.2.2
funcsigs==1.0.2
mock==2.0.0
pbr==1.10.0
Pillow==3.3.1
psycopg2-binary==2.8.6
pytz==2020.5
six==1.10.0
sorl-thumbnail==12.4.1
tablib==3.0.0
